Add files via upload
SussexBudgetProductions:

You are employed as a data scientist at SussexBudgetProductions. The last movie produced by the company was a comedy-action-thriller with a budget of £500K. It failed at the box office and only grossed £100K. In order to keep the production company in business, the next film needs to make money. In a senior leadership meeting, Beff Jezos (the CEO) decided that either a Romance movie or a Horror movie should be created next. As a member of the data science team, you have been tasked to produce some statistics and evidence to help decide which genre might be better, and to recommend a director and an actor to approach to design the movie around. In order to do this, you have been given an IMDB dataset Download IMDB datasetto explore and analyse - the dataset is a sample of all the movies ever created extracted from the IMDB website. This dataset was chosen in particular as it contains information about the budget, gross profit and it has the "IMDB scoreLinks to an external site."

The underlying assumption made by the senior leadership team is that a higher IMDB score implies a higher profit, so they are looking to create a movie that could achieve a high IMDB score. One can see that IMDB score is community driven, and so, as your manager, I am recommending that you think carefully about the number of reviews a movie has received (less reviews mean less certainty in the IMDB score).